+++ src/render/extract.rs
@@ -10,4 +10,4 @@ pub enum VelloExtractStep {
 
 /// A screenspace render target. We use a resizable fullscreen quad.
 #[derive(Component, Default, Clone, ExtractComponent)]
-pub struct SSRenderTarget(pub Handle<Image>);
+pub struct VelloRenderTarget(pub Handle<Image>);